+++ TypeSharp.Attributes/Enums/UnionAttribute.cs
@@ -1,4 +1,4 @@
-﻿namespace TypeSharp.Attributes.Attributes
+﻿namespace TypeSharp.Attributes.Enums
 {
     [AttributeUsage(AttributeTargets.Enum, AllowMultiple = false)]
     public class UnionAttribute : Attribute